Move decoders to own module
