+++ qiskit_ibm_runtime/decoders/noise_learner_v3.py
@@ -24,9 +24,9 @@ from ibm_quantum_schemas.noise_learner_v3.version_0_2 import (
     NoiseLearnerV3ResultsModel as NoiseLearnerV3ResultsModel_0_2,
 )
 
-from ..utils.result_decoder import ResultDecoder
-from .converters.version_0_1 import noise_learner_v3_result_from_0_1
-from .converters.version_0_2 import noise_learner_v3_result_from_0_2
+from .result_decoder import ResultDecoder
+from ..noise_learner_v3.converters.version_0_1 import noise_learner_v3_result_from_0_1
+from ..noise_learner_v3.converters.version_0_2 import noise_learner_v3_result_from_0_2
 
 if TYPE_CHECKING:
     from qiskit_ibm_runtime.noise_learner_v3.noise_learner_v3_result import NoiseLearnerV3Results
