+++ test/unit/quantum_program/test_result_decoders.py
@@ -33,8 +33,8 @@ from qiskit_ibm_runtime.quantum_program.quantum_program_result import (
     Metadata,
     QuantumProgramResult,
 )
-from qiskit_ibm_runtime.quantum_program.result_decoders import QuantumProgramResultDecoder
-from qiskit_ibm_runtime.quantum_program.executor_sampler import ExecutorSamplerResultDecoder
+from qiskit_ibm_runtime.decoders.quantum_program import QuantumProgramResultDecoder
+from qiskit_ibm_runtime.decoders.executor_sampler import ExecutorSamplerResultDecoder
 
 from ...ibm_test_case import IBMTestCase
